refactor: deduplicate provider and storage boilerplate
- Add default _create_empty_dataframe() to BaseProvider (canonical 7-column schema)
- Add default fetch_ohlcv_async() to BaseProvider (asyncio.to_thread wrapper)
- Remove duplicate _create_empty_dataframe from 6 providers (yahoo, twelve_data, fred, oanda, kalshi, polymarket)
- Remove duplicate fetch_ohlcv_async from 6 providers (yahoo, tiingo, polygon, finnhub, oanda, databento)
- Move shared storage methods to StorageBackend base class (get_metadata, _atomic_write, _update_metadata, _write_metadata_file)
- Extract _ensure_datetime to module-level function in update_manager.py (was duplicated in 3 classes)
- Fix EODHD/Tiingo tests that expected AttributeError from missing _create_empty_dataframe
