+++ src/ml4t/data/providers/oanda.py
@@ -326,48 +326,3 @@ class OandaProvider(BaseProvider):
             raise DataValidationError(
                 provider="oanda", message=f"Failed to transform data for {symbol}: {e}"
             )
-
-    def _create_empty_dataframe(self) -> pl.DataFrame:
-        """Create an empty DataFrame with the correct schema."""
-        return pl.DataFrame(
-            {
-                "timestamp": [],
-                "symbol": [],
-                "open": [],
-                "high": [],
-                "low": [],
-                "close": [],
-                "volume": [],
-            }
-        ).with_columns(
-            [
-                pl.col("timestamp").cast(pl.Datetime),
-                pl.col("symbol").cast(pl.String),
-                pl.col("open").cast(pl.Float64),
-                pl.col("high").cast(pl.Float64),
-                pl.col("low").cast(pl.Float64),
-                pl.col("close").cast(pl.Float64),
-                pl.col("volume").cast(pl.Float64),
-            ]
-        )
-
-    async def fetch_ohlcv_async(
-        self, symbol: str, start: str, end: str, frequency: str = "daily"
-    ) -> pl.DataFrame:
-        """Async fetch OHLCV data using thread pool.
-
-        Since the OANDA SDK is synchronous, this wraps the sync call
-        in asyncio.to_thread() to avoid blocking the event loop.
-
-        Args:
-            symbol: Forex pair (e.g., "EUR_USD")
-            start: Start date (YYYY-MM-DD)
-            end: End date (YYYY-MM-DD)
-            frequency: Data frequency
-
-        Returns:
-            Polars DataFrame with OHLCV data
-        """
-        import asyncio
-
-        return await asyncio.to_thread(self.fetch_ohlcv, symbol, start, end, frequency)
