+++ src/ml4t/data/providers/polygon.py
@@ -219,24 +219,3 @@ class PolygonProvider(BaseProvider):
             raise DataValidationError(
                 provider="polygon", message=f"Failed to transform data for {symbol}"
             ) from err
-
-    async def fetch_ohlcv_async(
-        self, symbol: str, start: str, end: str, frequency: str = "daily"
-    ) -> "pl.DataFrame":
-        """Async fetch OHLCV data using thread pool.
-
-        Since the Polygon SDK is synchronous, this wraps the sync call
-        in asyncio.to_thread() to avoid blocking the event loop.
-
-        Args:
-            symbol: Symbol to fetch
-            start: Start date (YYYY-MM-DD)
-            end: End date (YYYY-MM-DD)
-            frequency: Data frequency
-
-        Returns:
-            Polars DataFrame with OHLCV data
-        """
-        import asyncio
-
-        return await asyncio.to_thread(self.fetch_ohlcv, symbol, start, end, frequency)
