+++ src/ml4t/data/providers/tiingo.py
@@ -236,24 +236,3 @@ class TiingoProvider(BaseProvider):
             raise DataValidationError(
                 provider="tiingo", message=f"Failed to transform data for {symbol}"
             ) from err
-
-    async def fetch_ohlcv_async(
-        self, symbol: str, start: str, end: str, frequency: str = "daily"
-    ) -> "pl.DataFrame":
-        """Async fetch OHLCV data using thread pool.
-
-        Since the Tiingo SDK is synchronous, this wraps the sync call
-        in asyncio.to_thread() to avoid blocking the event loop.
-
-        Args:
-            symbol: Symbol to fetch
-            start: Start date (YYYY-MM-DD)
-            end: End date (YYYY-MM-DD)
-            frequency: Data frequency
-
-        Returns:
-            Polars DataFrame with OHLCV data
-        """
-        import asyncio
-
-        return await asyncio.to_thread(self.fetch_ohlcv, symbol, start, end, frequency)
